+++ ui/changelog.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## v0.3.1
+- skip athletes with incomplete name or result
+- print location of extracted file
+- remove debug messages from log
+- bugfixes:
+    - fix extracting people from database without ID
+    - do not extract LEV data to club column
+    - remove team id if incomplete
+
+## v0.3.0
+- add UI for extracting competition results from FSM MySQL database
+- update masterData for DEU officials and clubs
+- ignore rows in Meldeformular if no name is given
+- bugfixes:
+    - copy flags to correct fsm folder
+    - fix nation for M. Derpa
+
 ## v0.2.3
 - add flags for website
 - change default category to advanced novice